Prevent built-ins being edited in UI, to avoid confusion.
Also, disable "Add as new type" button when built-ins are selected. This
was never a good design. Better would be to have a way to extend
built-ins, but that needs a way to properly identify the built-in items,
so I'll settle for making the UI better communicate current limitations
for now.